+++ data-verification-ui/src/app/layout/Shell.jsx
@@ -1,6 +1,7 @@
 import ModuleNav from "./ModuleNav";
 import SideNav from "./SideNav";
 import TerminalCommandBar from "../../components/TerminalCommandBar";
+import GlobalGateBadge from "../../components/GlobalGateBadge";
 
 export default function Shell({ children, hideModuleNav = false }) {
   return (
@@ -8,7 +9,9 @@ export default function Shell({ children, hideModuleNav = false }) {
       {!hideModuleNav ? <SideNav /> : null}
       <div className="flex min-h-0 flex-1 flex-col">
         {!hideModuleNav ? <ModuleNav /> : null}
-        {!hideModuleNav ? <TerminalCommandBar /> : null}
+        {!hideModuleNav ? (
+          <TerminalCommandBar trailing={<GlobalGateBadge />} />
+        ) : null}
         {children}
       </div>
     </div>